refactor: make task tool async-only, remove delegate/delegation tools
- task tool now uses TaskManager.start() for fire-and-forget execution
- Removed sync SessionPrompt.prompt() code path entirely
- Deleted delegate.ts, delegation-read.ts, delegation-list.ts
- Updated registry to use task_read and task_list instead
- Updated task.txt docs to explain async behavior
- Consistent naming with PR-55 (task, task_read, task_list)
