+++ packages/opencode/src/tool/registry.ts
@@ -6,9 +6,8 @@ import { GrepTool } from "./grep"
 import { BatchTool } from "./batch"
 import { ReadTool } from "./read"
 import { TaskTool } from "./task"
-import { DelegateTool } from "./delegate"
-import { DelegationReadTool } from "./delegation-read"
-import { DelegationListTool } from "./delegation-list"
+import { TaskReadTool } from "./task-read"
+import { TaskListTool } from "./task-list"
 import { TodoWriteTool, TodoReadTool } from "./todo"
 import { WebFetchTool } from "./webfetch"
 import { WriteTool } from "./write"
@@ -111,9 +110,8 @@ export namespace ToolRegistry {
       EditTool,
       WriteTool,
       TaskTool,
-      DelegateTool,
-      DelegationReadTool,
-      DelegationListTool,
+      TaskReadTool,
+      TaskListTool,
       WebFetchTool,
       TodoWriteTool,
       // TodoReadTool,
